requirements.txt
openai
python-dotenv
pyaudio
viam-robot
Pillow
